chore: [#188595636] move form context functions out of types.ts (prefactor)
