+++ web/src/components/data-fields/Industry.tsx
@@ -3,13 +3,13 @@ import { EssentialQuestionField } from "@/components/data-fields/EssentialQuesti
 import { HomeContractor } from "@/components/data-fields/HomeContractor";
 import { IndustryDropdown } from "@/components/data-fields/IndustryDropdown";
 import { ConfigType } from "@/contexts/configContext";
+import { FormContextFieldProps } from "@/contexts/formContext";
 import { ProfileDataContext } from "@/contexts/profileDataContext";
 import { ProfileFormContext } from "@/contexts/profileFormContext";
 import { useConfig } from "@/lib/data-hooks/useConfig";
 import { useFormContextFieldHelpers } from "@/lib/data-hooks/useFormContextFieldHelpers";
 import { EssentialQuestion, EssentialQuestions } from "@/lib/domain-logic/essentialQuestions";
 import { getProfileConfig } from "@/lib/domain-logic/getProfileConfig";
-import { FormContextFieldProps } from "@/lib/types/types";
 import { ReactElement, ReactNode, useContext } from "react";
 
 interface Props<T> extends FormContextFieldProps<T> {
