+++ web/src/components/data-fields/tax-id/TaxIdFormContext.ts
@@ -1,5 +1,4 @@
-import { createFormContext } from "@/contexts/formContext";
-import { createReducedFieldStates } from "@/lib/types/types";
+import { createFormContext, createReducedFieldStates } from "@/contexts/formContext";
 
 export const taxIdFormContextErrorMap = createReducedFieldStates(["taxId", "taxIdLocation"]);
 type ErrorMapType = typeof taxIdFormContextErrorMap;
