+++ web/src/components/field-labels/FieldLabelModal.tsx
@@ -2,8 +2,9 @@ import { Content } from "@/components/Content";
 import { ProfileDataContext } from "@/contexts/profileDataContext";
 import { useConfig } from "@/lib/data-hooks/useConfig";
 import { getProfileConfig } from "@/lib/domain-logic/getProfileConfig";
-import { ProfileContentField } from "@/lib/types/types";
 import { ReactElement, useContext } from "react";
+
+import { ProfileContentField } from "@/contexts/profileFormContext";
 
 interface FieldOverrides {
   header: string | undefined;
