+++ web/src/components/field-labels/FieldLabelOnboarding.tsx
@@ -4,8 +4,10 @@ import { Heading } from "@/components/njwds-extended/Heading";
 import { ProfileDataContext } from "@/contexts/profileDataContext";
 import { useConfig } from "@/lib/data-hooks/useConfig";
 import { getProfileConfig } from "@/lib/domain-logic/getProfileConfig";
-import { FlowType, ProfileContentField } from "@/lib/types/types";
+import { FlowType } from "@/lib/types/types";
 import { ReactElement, useContext } from "react";
+
+import { ProfileContentField } from "@/contexts/profileFormContext";
 
 interface Props {
   fieldName: ProfileContentField;
