+++ web/src/components/profile/ProfileLockedField.tsx
@@ -1,7 +1,8 @@
 import { FieldLabelProfile } from "@/components/field-labels/FieldLabelProfile";
 import { ProfileDataContext } from "@/contexts/profileDataContext";
-import { ProfileContentField } from "@/lib/types/types";
 import { ReactElement, useContext } from "react";
+
+import { ProfileContentField } from "@/contexts/profileFormContext";
 
 export interface Props {
   fieldName: ProfileContentField;
