+++ web/src/components/profile/ProfileOpportunitiesAlert.tsx
@@ -2,9 +2,10 @@ import { FieldEntryAlert } from "@/components/FieldEntryAlert";
 import { ProfileDataContext } from "@/contexts/profileDataContext";
 import { useConfig } from "@/lib/data-hooks/useConfig";
 import { getProfileConfig } from "@/lib/domain-logic/getProfileConfig";
-import { ProfileContentField } from "@/lib/types/types";
 import { getFieldsForProfile, isFieldAnswered } from "@businessnjgovnavigator/shared";
 import { ReactElement, useContext } from "react";
+
+import { ProfileContentField } from "@/contexts/profileFormContext";
 
 export const ProfileOpportunitiesAlert = (): ReactElement => {
   const { state } = useContext(ProfileDataContext);
