+++ web/src/lib/cms/previews/ProfileFieldsPreview.tsx
@@ -23,13 +23,16 @@ import { Heading } from "@/components/njwds-extended/Heading";
 import { ProfileDocuments } from "@/components/profile/ProfileDocuments";
 import { LegalStructureRadio } from "@/components/tasks/business-structure/LegalStructureRadio";
 import { ConfigContext } from "@/contexts/configContext";
-import { ProfileFormContext } from "@/contexts/profileFormContext";
+import {
+  createProfileFieldErrorMap,
+  ProfileContentField,
+  ProfileFormContext,
+} from "@/contexts/profileFormContext";
 import { PreviewProps } from "@/lib/cms/helpers/previewHelpers";
 import { usePreviewConfig } from "@/lib/cms/helpers/usePreviewConfig";
 import { usePreviewRef } from "@/lib/cms/helpers/usePreviewRef";
 import { useFormContextHelper } from "@/lib/data-hooks/useFormContextHelper";
 import { EssentialQuestions } from "@/lib/domain-logic/essentialQuestions";
-import { createProfileFieldErrorMap, ProfileContentField } from "@/lib/types/types";
 import { businessStructureTaskId } from "@businessnjgovnavigator/shared/";
 import {
   IndustrySpecificData,
