+++ web/src/lib/domain-logic/essentialQuestions.ts
@@ -1,10 +1,11 @@
-import { ProfileContentField } from "@/lib/types/types";
 import {
   emptyIndustrySpecificData,
   Industry,
   IndustrySpecificData,
   LookupIndustryById,
 } from "@businessnjgovnavigator/shared";
+
+import { ProfileContentField } from "@/contexts/profileFormContext";
 
 export const getResetIndustrySpecificData = (
   industryId: string | undefined
